mini_crm/requirements.txt
Django>=4.2,<5.0
djangorestframework
python-dotenv